Refactored and fixed tests
