+++ evennia/contrib/django/__init__.py
@@ -1,0 +1,3 @@
+"""
+Intended to be a collecting folder for Django-specific contribs that do not have observable effects to players.
+"""
